+++ apps/ledger-live-mobile/src/screens/RequestAccount/02-SelectAccount.tsx
@@ -4,7 +4,7 @@ import { Trans } from "react-i18next";
 import type { Account, AccountLike, SubAccount } from "@ledgerhq/types-live";
 import { useSelector } from "react-redux";
 import { CompositeScreenProps, useTheme } from "@react-navigation/native";
-import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
+import { CryptoCurrency, CryptoOrTokenCurrency, TokenCurrency } from "@ledgerhq/types-cryptoassets";
 import { useGetAccountIds } from "@ledgerhq/live-common/wallet-api/react";
 import { accountsByCryptoCurrencyScreenSelector } from "~/reducers/accounts";
 import { TrackScreen } from "~/analytics";
@@ -23,6 +23,9 @@ import type {
 import { RequestAccountNavigatorParamList } from "~/components/RootNavigator/types/RequestAccountNavigator";
 import { BaseNavigatorStackParamList } from "~/components/RootNavigator/types/BaseNavigator";
 import { Flex } from "@ledgerhq/native-ui";
+import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
+import { useGroupedCurrenciesByProvider } from "@ledgerhq/live-common/deposit/index";
+import { LoadingBasedGroupedCurrencies } from "@ledgerhq/live-common/deposit/type";
 
 const SEARCH_KEYS = [
   "name",
@@ -99,6 +102,7 @@ function SelectAccount({ navigation, route }: Props) {
   const { accounts$, currency, allowAddAccount, onSuccess } = route.params;
   const accountIds = useGetAccountIds(accounts$);
   const accounts = useSelector(accountsByCryptoCurrencyScreenSelector(currency, accountIds));
+  const llmNetworkBasedAddAccountFlow = useFeature("llmNetworkBasedAddAccountFlow");
   const onSelect = useCallback(
     (account: AccountLike, parentAccount?: Account) => {
       onSuccess && onSuccess(account, parentAccount);
@@ -114,27 +118,76 @@ function SelectAccount({ navigation, route }: Props) {
     [onSelect],
   );
 
+  const { result, loadingStatus: providersLoadingStatus } = useGroupedCurrenciesByProvider(
+    true,
+  ) as LoadingBasedGroupedCurrencies;
+
+  const { currenciesByProvider } = result;
+
+  const provider = useMemo(
+    () =>
+      currenciesByProvider.find(elem =>
+        elem.currenciesByNetwork.some(
+          currencyByNetwork =>
+            (currencyByNetwork as CryptoCurrency | TokenCurrency).id === currency.id,
+        ),
+      ),
+    [currenciesByProvider, currency],
+  );
+
   const onAddAccount = useCallback(() => {
-    navigation.navigate(NavigatorName.RequestAccountsAddAccounts, {
-      screen: ScreenName.AddAccountsSelectDevice,
-      params: {
-        currency: currency as CryptoOrTokenCurrency,
-        onSuccess: () => navigation.navigate(ScreenName.RequestAccountsSelectAccount, route.params),
-      },
-    });
-  }, [currency, navigation, route.params]);
+    if (llmNetworkBasedAddAccountFlow?.enabled) {
+      if (provider && provider?.currenciesByNetwork.length > 1) {
+        navigation.navigate(NavigatorName.AssetSelection, {
+          screen: ScreenName.SelectNetwork,
+          params: {
+            currency: currency.id,
+            context: "addAccounts",
+            onSuccess: () =>
+              navigation.navigate(ScreenName.RequestAccountsSelectAccount, route.params),
+          },
+        });
+      } else {
+        navigation.navigate(NavigatorName.DeviceSelection, {
+          screen: ScreenName.SelectDevice,
+          params: {
+            currency: currency as CryptoCurrency,
+            context: "addAccounts",
+            onSuccess: () =>
+              navigation.navigate(ScreenName.RequestAccountsSelectAccount, route.params),
+          },
+        });
+      }
+    } else {
+      navigation.navigate(NavigatorName.RequestAccountsAddAccounts, {
+        screen: ScreenName.AddAccountsSelectDevice,
+        params: {
+          currency: currency as CryptoOrTokenCurrency,
+          onSuccess: () =>
+            navigation.navigate(ScreenName.RequestAccountsSelectAccount, route.params),
+        },
+      });
+    }
+  }, [currency, navigation, route.params, llmNetworkBasedAddAccountFlow?.enabled, provider]);
 
   const renderFooter = useCallback(
     () =>
       allowAddAccount ? (
         <View style={styles.buttonContainer}>
           <Button
+            disabled={
+              llmNetworkBasedAddAccountFlow?.enabled && providersLoadingStatus === "pending"
+            }
             containerStyle={styles.button}
             event="ExchangeStartBuyFlow"
             type="primary"
             title={
               <Trans
-                i18nKey="requestAccount.selectAccount.addAccount"
+                i18nKey={
+                  llmNetworkBasedAddAccountFlow?.enabled
+                    ? "addAccounts.addNewOrExisting"
+                    : "requestAccount.selectAccount.addAccount"
+                }
                 values={{
                   currency: currency.name,
                 }}
@@ -144,7 +197,13 @@ function SelectAccount({ navigation, route }: Props) {
           />
         </View>
       ) : null,
-    [allowAddAccount, currency.name, onAddAccount],
+    [
+      allowAddAccount,
+      currency.name,
+      onAddAccount,
+      llmNetworkBasedAddAccountFlow?.enabled,
+      providersLoadingStatus,
+    ],
   );
 
   const renderList = useCallback(