+++ libs/ui/examples/next.js/CHANGELOG.md
@@ -1,11 +1,18 @@
 # next.js-example
 
-## 0.1.12-nightly.0
+## 0.1.12
 
 ### Patch Changes
 
-- Updated dependencies []:
-  - @ledgerhq/react-ui@0.17.1-nightly.0
+- Updated dependencies [[`2a9e2de`](https://github.com/LedgerHQ/ledger-live/commit/2a9e2de3d14958123f702e4bde93eeb52967cddd)]:
+  - @ledgerhq/react-ui@0.18.0
+
+## 0.1.12-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`2a9e2de`](https://github.com/LedgerHQ/ledger-live/commit/2a9e2de3d14958123f702e4bde93eeb52967cddd)]:
+  - @ledgerhq/react-ui@0.18.0-next.0
 
 ## 0.1.11